requirements.txt
asgiref==3.2.10
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.8
defusedxml==0.7.1
distlib==0.3.3
Django==3.0.8
django-allauth==0.46.0
django-environ==0.4.5
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
filelock==3.4.0
gunicorn==20.0.4
idna==3.3
mysqlclient==2.1.0
oauthlib==3.1.1
platformdirs==2.4.0
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.3.1
urllib3==1.26.7
virtualenv==20.10.0
django-cors-headers==3.10.0